Add ability for content to specify a size.
CardView calculates the current size of the body, and passes it on to the CardSize.
CardSize passes that on to the CardContent.
If the content has a specified size, CardContent uses that as a fraction of the passed body size and forwards it to the relevant media component.

Create CardMetiaError component.